Look at Description
Added Scripts for a Game Manager, PlayerPrefs, RunInfo and Scrolling Image.

Scrolling Image is used in the case we want to use a unique background. It loops through the image.

Added code for functions and "psuedo code" for functions.

RunInfo script is used to close out the runinfo using the backbutton (not finished).

PlayerPrefs manages the PlayerPreferences with setters, getters, and a clear function.

GameManager is blank for now.