Changes to ShopManager for Sprites and functionality.
Fixed the shop such that the cards are not always the same type (ie. always two mentors or always two textbooks).

Added code to grab the card sprites on start.

Changed probability of Textbook and CardBuffs to be 15% respectively. (Can change to variables if we have vouchers that increase chances).

Added code that brings up card details for all three types of cards.

Added in a FadeIn and FadeOut function for cleaner animations.

NextRound now fades out the shop UI when pressed.