Added main UI into the main "Balatro-Feel scene".
Main and Shop UI are set to alpha 0, not interactable and not blocking raycasts. To see them, set these to active and or 1.
I moved some of the UI to better fit in but it is NOT final and will be changed later on, primarily where the deck is as it occupies too much of the center screen.
Some features are subject to change such as when next round is clicked, the title of shop will be changed depending on what the Ante is (etc.).
If anything is wrong, please let me know before changing anything regarding the Shop.