Changes in scene for UI, fade script made and TitleManager made.
In Scene for UI, added inclusions for the the different title changes that will occur during gameplay (Choose an Round, Shop Title, Round Title).

FadeScript made to allow different classes to call the FadeIn and FadeOut functions as needed. Simply pass in the Canvas Group.

TitleManager script created. Not fully functioning but will be used to change between the different titles stated before. General functions are within the script but needed to be fleshed out and tested.