Changes in Scene for UI, added code for TitleManager and started working on script to show current hand type.
Added to text for the current hand in the scene for UI. Attached the objects to the script for Title Manager.

Title Manager additions to setRoundTitle. Not all variables for the round are created so it will be updated in the future to have the actual names, symbols and scores appear. Added an instance so that it can be easily accessible. Removed some useless/filler code from functions.
CurrentHandManager is used to update the current hand's text in the UI. This includes the hand name and its level. This is simply to add the script and has no meaningful code as of right now.