Changes mostly to CurrentHandManager
Worked on the logic for detecting the different hands that can be played. Essential logic is implemented, though changes may need to be made depending on textbooks and whether we want to change precedence depending on score given from specific levels of hands.
Changed runinfo in UIManager to be non-interactable when not active on screen.