Changes to AudioManager, ShopManager in addition to new scripts.
Fixed function in AudioManager to correctly transition tracks to shop.

Changed how a function was called for transitionmanager to correctly switch between screens.

DefeatManager sets the text for when the player losses.

EndOfRoundManager sets the text for when the player wins in addition to the cashout function which when the button is pressed, updates the player's moneyCount and transitions to the shopScreen and audio.

Updates to Scrolling Image to rotate.

TransitionManager is used as a centralized script to switch between the different screens and update the different music tracks as needed.

Added screens for the different UIs in the Scene-For-Creating-ShopUI.