Additions to ShakeScreen, ScoringManager, and CurrentHandManager.
CurrentHandManager return strings have been changed to match their names within the Textbook table. This helps when passing them into the scoring manager and streamlines the process.

ShakeScreen duration reduced.

ScoringManager now has additional functions. This includes gathering the different required variables, a function to go through each card and add its mult/chips to the current number and update it in the UI. Changes must be made to both test if the script works in addition to Joker effects.