Check Description. Change pertain to Victory screen.
Created Victory Object in new scene.
VictoryManager holds variables that are displayed in the victory screen and sets the text.
Updates to ScoringManager and ShopManager to update the different variables from Victory Manager.
Updates Scoring Manager to detect if the game was won during the 8th ante and 3rd round and transition to victory screen.
Victory screen function added in transition manager.